Move logic from DateConfig component into hook.
